Refactor template to separate packages, extract request handlers and add REST api
