+++ template/faas-flow/config/consul_dc.go
@@ -1,0 +1,13 @@
+package config
+
+import (
+	"os"
+)
+
+func ConsulDC() string {
+	val := os.Getenv("consul_dc")
+	if len(val) == 0 {
+		val = "dc1"
+	}
+	return val
+}
