+++ template/faas-flow/config/read_timeout.go
@@ -1,0 +1,10 @@
+package config
+
+import (
+	"os"
+	"time"
+)
+
+func ReadTimeout() time.Duration {
+	return parseIntOrDurationValue(os.Getenv("read_timeout"), 10*time.Second)
+}
