+++ template/faas-flow/config/trace_server.go
@@ -1,0 +1,14 @@
+package config
+
+import (
+	"os"
+)
+
+// TraceServer get the traceserver address
+func TraceServer() string {
+	traceServer := os.Getenv("trace_server")
+	if traceServer == "" {
+		traceServer = "jaeger.faasflow:5775"
+	}
+	return traceServer
+}
