+++ template/faas-flow/config/write_timeout.go
@@ -1,0 +1,10 @@
+package config
+
+import (
+	"os"
+	"time"
+)
+
+func WriteTimeout() time.Duration {
+	return parseIntOrDurationValue(os.Getenv("write_timeout"), 10*time.Second)
+}
