+++ template/faas-flow/executor/executor.go
@@ -1,0 +1,8 @@
+package executor
+
+import "github.com/s8sg/faas-flow/sdk"
+
+var (
+	stateStore sdk.StateStore
+	dataStore  sdk.DataStore
+)
