+++ template/faas-flow/executor/handle_error.go
@@ -1,0 +1,13 @@
+package executor
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func handleError(w http.ResponseWriter, message string) {
+	errorStr := fmt.Sprintf("[ Failed ] %v\n", message)
+	fmt.Printf(errorStr)
+	w.WriteHeader(http.StatusInternalServerError)
+	w.Write([]byte(errorStr))
+}
