+++ template/faas-flow/executor/init_request_tracer.go
@@ -1,0 +1,46 @@
+package executor
+
+import (
+	"fmt"
+	"time"
+
+	hconfig "handler/config"
+
+	"github.com/opentracing/opentracing-go"
+	"github.com/uber/jaeger-client-go"
+	"github.com/uber/jaeger-client-go/config"
+)
+
+// initRequestTracer init global trace with configuration
+func initRequestTracer(flowName string) (*traceHandler, error) {
+	tracerObj := &traceHandler{}
+
+	agentPort := hconfig.TraceServer()
+
+	cfg := config.Configuration{
+		ServiceName: flowName,
+		Sampler: &config.SamplerConfig{
+			Type:  "const",
+			Param: 1,
+		},
+		Reporter: &config.ReporterConfig{
+			LogSpans:            true,
+			BufferFlushInterval: 1 * time.Second,
+			LocalAgentHostPort:  agentPort,
+		},
+	}
+
+	opentracer, traceCloser, err := cfg.NewTracer(
+		config.Logger(jaeger.StdLogger),
+	)
+	if err != nil {
+		return nil, fmt.Errorf("failed to init tracer, error %v", err.Error())
+	}
+
+	tracerObj.closer = traceCloser
+	tracerObj.tracer = opentracer
+	tracerObj.nodeSpans = make(map[string]opentracing.Span)
+	tracerObj.operationSpans = make(map[string]map[string]opentracing.Span)
+
+	return tracerObj, nil
+}
