+++ template/faas-flow/executor/new_request_handler_wrapper.go
@@ -1,0 +1,29 @@
+package executor
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/julienschmidt/httprouter"
+	"github.com/s8sg/faas-flow/sdk/executor"
+)
+
+func newRequestHandlerWrapper(handler func(http.ResponseWriter, *http.Request, string, executor.Executor) ([]byte, error)) func(http.ResponseWriter, *http.Request, httprouter.Params) {
+	return func(w http.ResponseWriter, req *http.Request, params httprouter.Params) {
+		id := params.ByName("id")
+
+		ex, err := newRequestExecutor(req)
+		if err != nil {
+			handleError(w, fmt.Sprintf("failed to execute request "+id))
+			return
+		}
+
+		body, err := handler(w, req, id, ex)
+		if err != nil {
+			handleError(w, fmt.Sprintf("request failed to be processed. "+err.Error()))
+		}
+
+		w.WriteHeader(http.StatusOK)
+		w.Write(body)
+	}
+}
