+++ template/faas-flow/go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/codahale/hdrhistogram v0.0.0-20161010025455-3a0bb77429bd // indirect
 	github.com/go-ini/ini v1.56.0 // indirect
 	github.com/hashicorp/consul/api v1.4.0 // indirect
+	github.com/julienschmidt/httprouter v1.3.0
 	github.com/minio/minio-go v6.0.14+incompatible // indirect
 	github.com/opentracing/opentracing-go v1.1.0
 	github.com/pkg/errors v0.9.1 // indirect
