+++ template/faas-flow/log/std_out_logger.go
@@ -1,0 +1,17 @@
+package log
+
+import (
+	"fmt"
+)
+
+// implements faasflow.Logger
+type StdOutLogger struct{}
+
+func (l *StdOutLogger) Configure(flowName string, requestId string) {}
+
+func (l *StdOutLogger) Init() error {
+	return nil
+}
+func (l *StdOutLogger) Log(str string) {
+	fmt.Print(str)
+}
